pyproject.toml
[project]
name = "comarch-woocomerence-sync"
version = "1.0"
description = ""
authors = [
    {name = "maxthesyntax",email = "maks@x-net.pl"}
]
requires-python = ">=3.14"
dependencies = [
    "woocommerce (>=3.0.0,<4.0.0)",
    "python-dotenv (>=1.2.1,<2.0.0)",
    "pyodbc (>=5.3.0,<6.0.0)",
    "wordpress-api (>=1.2.9,<2.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
